fix(bridge): split git api commands into part file
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
