+++ .sisyphus/notepads/merge-main-port-branch-selection/learnings.md
@@ -51,3 +51,4 @@
 - Verification evidence for the no-op result: focused `module_core` tests for project API, repository, session service, and new-session cubit passed; `dart analyze` in `mobile/module_core` passed cleanly; `lsp_diagnostics` reported no issues on the touched core files.
 - Branch parsing should not treat a remote name like `origin` as a pseudo-ref filter; the real guard is excluding `origin/HEAD`-style refs while keeping a local branch named `origin` selectable.
 - Final-review cleanup can stay layered by moving archive/unarchive orchestration into `SessionArchiveStatusService`, routing cleanup through a service-layer helper, and pulling session worktree creation into `WorktreeSessionPreparationService`; that kept `update_session_archive_status_handler.dart` and `worktree_service.dart` comfortably under the file-size cap without changing behavior.
+- `GitCliApi` can stay a single git wrapper while its command methods live in a `part` file; that trims the main API wrapper under the file cap without splitting the external tool abstraction by use-case.
